+++ app/src/main/java/com/github/nisrulz/samplezentone/screen/main/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.github.nisrulz.samplezentone.ui.screen.main
+package com.github.nisrulz.samplezentone.screen.main
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -12,6 +12,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.PreviewLightDark
+import com.github.nisrulz.samplezentone.screen.main.ui.LoadingScreen
+import com.github.nisrulz.samplezentone.screen.main.ui.SuccessScreen
 import com.github.nisrulz.samplezentone.ui.theme.AppTheme
